+++ CHANGES.md
@@ -12,6 +12,7 @@ CHANGES
 * Fix: Comment body doesn't break long terms
 * Fix: lsx plugin lists up pages that hit by forward match wrongly
     * Introduced by 3.0.4
+* Support: Multilingualize React components with i18next
 * Support: Upgrade libs
     * elasticsearch
     * googleapis